add node to size_t

+++ 0x12-singly_linked_lists/2-add_node.c
@@ -11,7 +11,7 @@
 list_t *add_node(list_t **head, const char *str)
 {
 	list_t *newNode;
-	size_t length = 0;
+	unsigned int length = 0;
 
 	newNode = malloc(sizeof(list_t));
 	if (newNode == NULL)
